Move `FeedLoaderPresentationAdapter` to a separate file
